worklog/requirements.txt
fastapi>=0.100.0,<0.101.0
pydantic>=2.4.2,<3.0.0
pydantic-settings>=2.0.3,<3.0.0
uvicorn>=0.15.0,<0.16.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0,<2.0.0
pymysql>=1.0.2,<2.0.0
requests>=2.26.0,<3.0.0
python-dotenv>=0.21.0,<0.22.0
redis>=5.0.0,<6.0.0
python-magic>=0.4.27  # 用于文件类型检测
bcrypt>=4.0.1,<5.0.0
SpeechRecognition>=3.10.0  # 用于语音识别
jinja2>=3.0.0,<4.0.0  # 用于消息模板渲染
fastapi-mail>=1.4.1,<2.0.0  # 用于邮件发送
pydantic[email]>=2.0.0,<3.0.0  # 用于邮箱验证
pytest>=7.0.0,<8.0.0  # 用于测试
pytest-asyncio>=0.21.0,<1.0.0  # 用于异步测试
httpx>=0.24.0,<1.0.0  # 用于HTTP客户端测试
websockets>=11.0.0,<12.0.0  # 用于WebSocket支持
aiofiles>=23.0.0,<25.0.0  # 用于异步文件操作
celery>=5.3.0,<6.0.0  # 用于异步任务队列
schedule>=1.2.0,<2.0.0  # 用于定时任务
APScheduler>=3.10.0,<4.0.0  # 用于高级定时任务调度 